# lib/api-zod/src/generated/types/dailyTrend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Environmental Impact Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyTrend {
  date: Date;
  waterMl: number;
  co2G: number;
  energyWh: number;
  sessions: number;
  queries: number;
}