# lib/api-zod/src/generated/types/gridIntensity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Environmental Impact Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { GridIntensitySource } from './gridIntensitySource';

export interface GridIntensity {
  /** Grid region ID (matches the grid selector) */
  id: string;
  label: string;
  /** Carbon intensity in gCO2/kWh — live when source is "live", otherwise the static fallback */
  intensityGPerKwh: number;
  /** Whether the intensity came from the live EIA feed or the static curated average */
  source: GridIntensitySource;
  /**
     * Timestamp of the live reading (UTC hour), null for static values
     * @nullable
     */
  asOf?: Date | null;
}