# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Environmental Impact Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionInputActivityType } from './sessionInputActivityType';
import type { SessionInputComplexity } from './sessionInputComplexity';

export interface SessionInput {
  toolId: number;
  activityType: SessionInputActivityType;
  /** @minimum 1 */
  durationMinutes: number;
  complexity: SessionInputComplexity;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  sessionDate?: Date | null;
}